spec: move some per-library keys out to become per-package
This seems to be how current packages are described. Also 'licenses' is
renamed 'license' to reflect what chibi implements.

License in the original spec is about one license. However it makes
sense to define it as a list instead for dual-licensed
packages. (macduffie json) is currently defined with two as an
example.

Although that one seems wrong (lack of validation?), it's defined as

    (license Expat (MIT))